+++ src/theory/quantifiers/quant_rep_bound_ext.cpp
@@ -36,7 +36,7 @@ QRepBoundExt::QRepBoundExt(Env& env,
 }
 
 RsiEnumType QRepBoundExt::setBound(Node owner,
-                                   unsigned i,
+                                   size_t i,
                                    std::vector<Node>& elements)
 {
   // builtin: check if it is bound by bounded integer module
@@ -57,7 +57,7 @@ RsiEnumType QRepBoundExt::setBound(Node owner,
 
 bool QRepBoundExt::resetIndex(RepSetIterator* rsi,
                               Node owner,
-                              unsigned i,
+                              size_t i,
                               bool initial,
                               std::vector<Node>& elements)
 {
@@ -79,7 +79,7 @@ bool QRepBoundExt::initializeRepresentativesForType(TypeNode tn)
   return d_model->initializeRepresentativesForType(tn);
 }
 
-bool QRepBoundExt::getVariableOrder(Node owner, std::vector<unsigned>& varOrder)
+bool QRepBoundExt::getVariableOrder(Node owner, std::vector<size_t>& varOrder)
 {
   // must set a variable index order based on bounded integers
   if (owner.getKind() != FORALL)
