+++ src/theory/quantifiers/quant_rep_bound_ext.h
@@ -50,18 +50,18 @@ class QRepBoundExt : public RepBoundExt
   virtual ~QRepBoundExt() {}
   /** set bound */
   RsiEnumType setBound(Node owner,
-                       unsigned i,
+                       size_t i,
                        std::vector<Node>& elements) override;
   /** reset index */
   bool resetIndex(RepSetIterator* rsi,
                   Node owner,
-                  unsigned i,
+                  size_t i,
                   bool initial,
                   std::vector<Node>& elements) override;
   /** initialize representative set for type */
   bool initializeRepresentativesForType(TypeNode tn) override;
   /** get variable order */
-  bool getVariableOrder(Node owner, std::vector<unsigned>& varOrder) override;
+  bool getVariableOrder(Node owner, std::vector<size_t>& varOrder) override;
 
  private:
   /** Reference to the quantifiers bound inference */
@@ -69,7 +69,7 @@ class QRepBoundExt : public RepBoundExt
   /** Pointer to the quantifiers model */
   FirstOrderModel* d_model;
   /** indices that are bound integer enumeration */
-  std::map<unsigned, bool> d_bound_int;
+  std::map<size_t, bool> d_bound_int;
   /** An instantiation match */
   InstMatch d_instMatch;
 };
